WiP - separating lib and bin
